Moves Shape::Concept & Shape::Model out...
And, into detail::ShapeConcept & detail::ShapeModel respectively.
